+++ components/Loading/Loading.tsx
@@ -1,0 +1,11 @@
+import styles from "./Loading.module.scss";
+
+const Loading: React.FC = () => {
+    return (
+        <div className={styles.loading}>
+            <h2>Loading...</h2>
+        </div>
+    );
+};
+
+export default Loading;
