+++ components/Registration/Registration.module.scss
@@ -24,18 +24,6 @@
     }
 }
 
-.loading {
-    position: fixed;
-    inset: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0, 0, 0, 0.5);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 1000;
-}
-
 .contentWrapper {
     margin-top: 126px;
     display: flex;
